requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djoser==2.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
django-cors-headers==4.4.0
